Makefile: disable GETTEXT when 'po' is missing
As an effort to promote "dogfooding" the sparse-checkout feature within
the Git codebase, it is helpful to explore which portions of the
codebase are optional.

The NO_GETTEXT build variable can disable the translation libraries.
The test suite then uses the NO_GETTEXT environment variable to disable
the GETTEXT prerequisite for some tests.

The 'po' directory contains translations for strings in the Git
codebase. While this stores extremely important data for Git users, the
data is not typically used by Git contributors in their daily work.
Thus, it could be removed from the working directory using
sparse-checkout.

However, doing such a removal causes some tests to fail. Part of the
failures are related to the GETTEXT prerequisite being enabled when it
probably should not be. The other part is that some tests in
t0200-gettext-basic.sh depend on the existence of files in the po
directory.

In test-lib, disable the GETTEXT prerequisite when the po directory does
not exist, then add the GETTEXT prerequisite to these tests that use
files in that directory.

In Makefile, define NO_GETTEXT when the po directory does not exist.
This is necessary for "make install" to work correctly.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>