run_auto_gc(): extract a reusable helper
Earlier we taught "git fetch --quiet" to pass the "--quiet" option
down to "gc --auto".  This, however, is not limited to "fetch":

    $ git grep -e 'gc.*--auto' \*.c

finds hits in "am", "commit", "merge", and "rebase".

As a preparatory step, let's introduce a helper function
run_auto_gc(), that the caller can pass a boolean "quiet".

Signed-off-by: Junio C Hamano <gitster@pobox.com>