reftable doc: use link: and urlencode to avoid dead links
Showing as broken links in the html version (when using asciidoc 8.6.10)
because some of the characters in the url need encoding.

Change them to use a explicit "link" and URL encode the problematic
charecters (ex: +, =, @).  While at it, change the base URL to use
lore.kernel.org as has been suggested since 56e6c16394 (Merge branch
'dl/lore-is-the-archive', 2019-12-06)

Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>
Reviewed-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>