t3200: avoid variations of the `master` branch name
To avoid branch names with a loaded history, we already started to avoid
using the name "master" in a couple instances.

The `t3200-branch.sh` script uses variations of this name for branches
other than the default one. So let's change those names, as
"lowest-hanging fruits" in the effort to use more inclusive naming
throughout Git's source code.

In this particular instance, this requires a couple of non-trivial
adjustments, as the aligned output depends on the maximum length of the
displayed branches (which we now changed), and also on the alphabetical
order (which we now changed, too).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>