tests: avoid variations of the `master` branch name
The term `master` has a loaded history that serves as a constant
reminder of racial injustice. The Git project has no desire to
perpetuate this and already started avoiding it.

The test suite uses variations of this name for branches other than the
default one. Apart from t3200, where we just addressed this in the
previous commit, those instances can be renamed in an automated manner
because they do not require any changes outside of the test script, so
let's do that.

Seeing as `main` seems to be materializing as the favorite default
branch name of many Open Source projects, we use that name in our
replacements, too.

This trick was performed by this (GNU) sed invocation:

	$ sed -i 's/master\([a-z0-9]\)/main\1/g' t/t*.sh

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>