refs: update delete_ref() to take a struct repository
There are a number of places where we delete pseudo refs by calling
unlink() on the pseduo ref path rather than using the refs api. The
assumption that these refs are always files will not valid once
repositories start storing their refs in a reftable. As the current code
is already aware of which repository the ref is in update delete_ref()
to take a struct repository* so that we can safely change the unlink()
calls to delete_ref() later. It would be possible to change the unlink()
to use refs_delete_ref(get_main_ref_store(r), ...) but that gets quite
cumbersome.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>